(refactor) landing page
