+++ client/assets/stylesheet.css
@@ -1,11 +1,26 @@
-/*
- * Sources:
- * main.css https://github.com/callemall/material-ui/blob/master/docs/src/www/css/main.css
- * mui-github-markdown.css https://github.com/callemall/material-ui/blob/master/docs/src/app/components/mui-github-markdown.css
- */
+@import url("reset.css");
+
+* {
+  box-sizing: border-box;
+}
+
+html, body {
+  height: 100%;
+}
+
+body {
+  text-size-adjust: 100%;
+  color: #333;
+  font-family: "Roboto", sans-serif;
+  font-size: 1.125em;
+  line-height: 1.5;
+  word-wrap: break-word;
+  -webkit-text-size-adjust: 100%;
+  -webkit-font-smoothing: antialiased;
+}
 
 a {
-  color: #ff4081;
+  color: #3498DB; /* primary color */
   text-decoration: none;
 }
 
@@ -13,86 +28,49 @@ a:hover {
   text-decoration: underline;
 }
 
-html {
-  font-family: 'Roboto', sans-serif;
-  -webkit-font-smoothing: antialiased;
+.jumbotron {
+  background: #2980B9; /* primary color: dark */
+  color: #FFF;
+  height: 100vh;
+  position: relative;
+  text-align: center;
+  overflow: hidden;
 }
 
-body, h1, h2, h3, h4, h5, h6 {
-  margin: 0;
-}
+  .jumbotron h1 {
+    font-size: 2.369rem;
+    font-weight: lighter;
+  }
 
+  .jumbotron h2 {
+    font-size: 1.333rem;
+    font-weight: lighter;
+    margin-bottom: 1rem;
+  }
+
+.center {
+  max-width: 25em;
+  margin: 0 auto;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
 
 footer {
-  position: absolute;
-  right: 20px;
-  bottom: 20px;
-  width: 100%;
-  color: #AAAAAA;
-  text-align: right;
+  color: #AAA;
+  font-size: .75rem;
+  text-align: center;
+  padding: 1rem;
 }
 
 .github {
-  position: absolute;
-  top: 20px;
-  left: 20px;
-  width: 32px;
-  height: 32px;
-}
-
-.landing-outer {
-  display: table;
-  position: absolute;
-  height: 100%;
-  width: 100%;
-}
-
-.landing-inner {
-  display: table-cell;
-  vertical-align: middle;
-  text-align: center;
+  display: inline;
+  width: 32px !important;
+  height: 32px !important;
 }
 
 .logo {
-  width: 160px;
-}
-
-.floatingActionButton {
-  position: fixed;
-  bottom: 20px;
-  left: 20px;
-  z-index: 3;
-}
-
-.copy {
-  width: 250px;
-  height: 100px;
-  padding: 10px 0 0 10px;
-}
-
-.link {
-  padding-top: 15px;
-}
-
-input.url {
-  width: 200px;
-  font-size: 10pt;
-  height: 26px;
-  margin-right: 10px;
-  margin-bottom: 15px;
-}
-
-.copybtn {
-  height: 31px;
-  position: absolute;
-}
-
-.copybtn:hover {
-  border-color: rgba(27,31,35,0.35);
-}
-
-.clipboard {
-  width: 15px;
-  height: 20px;
-  padding-top: 3px;
+  width: 5.3rem !important;
+  height: 5.3rem !important;
 }
